phase2-vision-system/aether-console/requirements.txt
PyQt6>=6.7
opencv-python>=4.10
numpy>=1.26
requests>=2.32
ultralytics>=8.3
torch>=2.2
pyserial>=3.5
esptool>=4.8
pyinstaller>=6.10
